only delete problems if the file exists
git-svn-id: https://svn.objectstyle.org/repos/woproject/trunk@4741 930f2ab1-3212-0410-a8ea-d1eb4cc50a93

+++ wolips/core/plugins/org.objectstyle.wolips.wodclipse.core/java/org/objectstyle/wolips/wodclipse/core/util/WodModelUtils.java
@@ -90,7 +90,9 @@ public class WodModelUtils {
   
   public static void deleteWodProblems(IFile wodFile) {
     try {
-      wodFile.deleteMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE);
+      if (wodFile.exists()) {
+        wodFile.deleteMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE);
+      }
     }
     catch (CoreException e) {
       Activator.getDefault().debug(e);
